Move `check_value` & `check_single_vlaue` to `_optimize.py`
